20120302 Comsenz 核心产品 Discuz! X2.5 RC 发布
升级：
1. DiscuzX 程序  1.0, 1.5 2.0 ， 2.5Beta
1）备份数据库
2）建立文件夹 old，旧程序除了 data ，  config， uc_client, uc_server 目录以外的程序移动进入 old目录中
3） 上传 2.5 RC程序（压缩包中 upload 目录中的文件）
4）上传安装包 utility 目录中的 update.php 到论坛 install 目录，删除 install 目录中的index.php

执行 http://你的域名/论坛路径/install/update.php

参照提示进行升级即可。升级时间随着数据的大小和服务器性能而变， 依据测试，10G左右的论坛，从 2.0 升级值 2.5 大约需 1.5-2.5 个小时。如果你的版本更老一些，升级的时间消耗的会更多些。

升级完毕，进入后台，更新缓存，并测试功能。最后，你可以将old目录中所需要的东西移动回来

2. Discuz! 7.2 或以下版本的程序
1）首先参看我们的转换教程， 将程序转换到 X2.0
2）上传 2.5 RC程序
3）上传 utility 目录中的 update.php 到 install 目录，删除目录中的index.php

执行 http://你的域名/论坛路径/install/update.php

参照提示进行升级即可

+++ discuzx/upload/source/class/discuz/discuz_table.php
@@ -4,7 +4,7 @@
  *      [Discuz!] (C)2001-2099 Comsenz Inc.
  *      This is NOT a freeware, use is subject to license terms
  *
- *      $Id: discuz_table.php 27481 2012-02-02 04:28:46Z monkey $
+ *      $Id: discuz_table.php 28455 2012-03-01 04:55:11Z zhangguosheng $
  */
 
 if(!defined('IN_DISCUZ')) {
@@ -47,18 +47,24 @@ class discuz_table extends discuz_base
 	}
 
 	public function update($val, $data, $unbuffered = false, $low_priority = false) {
-		$this->checkpk();
-		$ret = DB::update($this->_table, $data, DB::field($this->_pk, $val), $unbuffered, $low_priority);
-		foreach((array)$val as $id) {
-			$this->update_cache($id, $data);
+		if(isset($val) && !empty($data) && is_array($data)) {
+			$this->checkpk();
+			$ret = DB::update($this->_table, $data, DB::field($this->_pk, $val), $unbuffered, $low_priority);
+			foreach((array)$val as $id) {
+				$this->update_cache($id, $data);
+			}
+			return $ret;
 		}
-		return $ret;
+		return !$unbuffered ? 0 : false;
 	}
 
 	public function delete($val, $unbuffered = false) {
-		$this->checkpk();
-		$ret = DB::delete($this->_table, DB::field($this->_pk, $val), null, $unbuffered);
-		$this->clear_cache($val);
+		$ret = false;
+		if(isset($val)) {
+			$this->checkpk();
+			$ret = DB::delete($this->_table, DB::field($this->_pk, $val), null, $unbuffered);
+			$this->clear_cache($val);
+		}
 		return $ret;
 	}
