+++ discuzx/upload/source/class/table/table_common_adminnote.php
@@ -4,7 +4,7 @@
  *      [Discuz!] (C)2001-2099 Comsenz Inc.
  *      This is NOT a freeware, use is subject to license terms
  *
- *      $Id: table_common_adminnote.php 27449 2012-02-01 05:32:35Z zhangguosheng $
+ *      $Id: table_common_adminnote.php 27800 2012-02-15 02:13:57Z svn_project_zhangjie $
  */
 
 if(!defined('IN_DISCUZ')) {
@@ -22,14 +22,23 @@ class table_common_adminnote extends discuz_table
 	}
 
 	public function delete($id, $admin = '') {
+		if(empty($id)) {
+			return false;
+		}
 		return DB::query('DELETE FROM %t WHERE '.DB::field('id', $id).' %i', array($this->_table, ($admin ? ' AND '.DB::field('admin', $admin) : '')));
 	}
 
 	public function fetch_all_by_access($access) {
-		 return DB::fetch_all('SELECT * FROM %t WHERE '.DB::field('access', $access).' ORDER BY dateline DESC', array($this->_table));
+		if(empty($access)) {
+			return array();
+		}
+		return DB::fetch_all('SELECT * FROM %t WHERE '.DB::field('access', $access).' ORDER BY dateline DESC', array($this->_table));
 	}
 
 	public function count_by_access($access) {
+		if(empty($access)) {
+			return 0;
+		}
 		return DB::result_first('SELECT COUNT(*) FROM %t WHERE '.DB::field('access', $access), array($this->_table));
 	}
